chore: fix circular deps, remove dead code
