+++ libs/wallet/src/connect-dialog/index.ts
@@ -1,2 +1,3 @@
 export * from './connect-dialog';
 export * from './view-as-dialog';
+export * from './vega-wallet-dialog-store';
