+++ libs/wallet/src/index.ts
@@ -1,14 +1,12 @@
+export * from './connectors';
 export * from './context';
 export * from './use-vega-wallet';
-export * from './connectors';
-export * from './use-vega-transaction';
 export * from './use-vega-transaction-manager';
 export * from './use-vega-transaction-store';
 export * from './use-vega-transaction-updater';
 export * from './use-transaction-result';
 export * from './use-eager-connect';
 export * from './manage-dialog';
-export * from './vega-transaction-dialog';
 export * from './provider';
 export * from './connect-dialog';
 export * from './utils';
