+++ libs/wallet/src/use-vega-transaction-store.spec.tsx
@@ -1,5 +1,7 @@
-import { useVegaTransactionStore } from './use-vega-transaction-store';
-import { VegaTxStatus } from './use-vega-transaction';
+import {
+  useVegaTransactionStore,
+  VegaTxStatus,
+} from './use-vega-transaction-store';
 import type { VegaStoredTxState } from './use-vega-transaction-store';
 import type {
   OrderAmendmentBody,
