--- libs/wallet/src/vega-transaction-dialog/index.ts
@@ -1,1 +1,0 @@
-export * from './vega-transaction-dialog';
